Refactor of library logging.
Add a configurable logger to dnstap Output and Input
types. This logger is used in the RunOutputLoop
and ReadInto loops to log errors and, in the case
of FrameStreamSockInput and FrameStreamSockOuput,
connection events. Logging is turned off by default.
Fatal error messages (log.Fatalf) replaced with
a log message (if configured) and a loop exit, which
may cause callers to hang but is preferable to
exploding the whole program.